pyproject.toml
[project]
name = "mas-dummy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "autogen-agentchat>=0.6.2",
    "autogen-ext[azure,openai]>=0.6.2",
    "python-dotenv>=1.1.1",
]
